some error fixed with temporary solution

+++ froentend/src/adminComponenets/MenuTable.jsx
@@ -1,39 +1,54 @@
+import React, { useEffect } from 'react';
 import {
   Avatar,
-  Box, Card, CardHeader, IconButton, Paper,
-  Table, TableBody, TableCell, TableContainer, TableHead, TableRow
-} from '@mui/material'
-import React, { useEffect } from 'react'
+  Box,
+  Card,
+  CardHeader,
+  IconButton,
+  Paper,
+  Table,
+  TableBody,
+  TableCell,
+  TableContainer,
+  TableHead,
+  TableRow,
+  CircularProgress
+} from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
 import { Delete, Edit } from '@mui/icons-material';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { deleteFoodAction, getMenuItemByRestaurantId } from '../state/Menu/Action'
+import { deleteFoodAction, getMenuItemByRestaurantId } from '../state/Menu/Action';
 
 const MenuTable = () => {
   const dispatch = useDispatch();
-  const jwt = localStorage.getItem("jwt");
-  const { restaurant, menu } = useSelector((store) => store);
-
   const navigate = useNavigate();
 
+  const jwt = localStorage.getItem("jwt");
+  const restaurant = useSelector((store) => store.restaurant);
+  const menu = useSelector((store) => store.menu);
+
+  const isLoading = menu.loading; // Optional: depends on your state structure
+
   useEffect(() => {
-    dispatch(getMenuItemByRestaurantId({
-      restaurantId : restaurant.usersRestaurant.id,
-      jwt,
-      // vegetarian : false,
-      // seasonal: false,
-      // nonveg:false
-    }))
-  }, [])
+    if (restaurant?.usersRestaurant?.id) {
+      dispatch(getMenuItemByRestaurantId({
+        restaurantId: restaurant.usersRestaurant.id,
+        jwt,
+      }));
+    }
+  }, [restaurant, jwt, dispatch]);
 
   const handleCreate = () => {
     navigate('addMenu');
   };
 
   const handleDeleteFood = (foodId) => {
-    dispatch(deleteFoodAction({foodId,jwt}))
-  }
+    const confirmDelete = window.confirm("Are you sure you want to delete this item?");
+    if (confirmDelete) {
+      dispatch(deleteFoodAction({ foodId, jwt }));
+    }
+  };
 
   return (
     <div className='mt-2 pt-1'>
@@ -57,36 +72,60 @@ const MenuTable = () => {
                 <TableCell align='right'>ID</TableCell>
                 <TableCell align="right">Image</TableCell>
                 <TableCell align="right">Name</TableCell>
-                <TableCell align='right'>Price</TableCell>
-                <TableCell align='right'>Description</TableCell>
-                <TableCell align="right">Availability</TableCell>
-                <TableCell align="right"></TableCell>
+                <TableCell align="right">Price</TableCell>
+                <TableCell align="right">Status</TableCell>
+                <TableCell align="right">Change</TableCell>
+                <TableCell align="right">Actions</TableCell>
               </TableRow>
             </TableHead>
 
             <TableBody>
-              {menu.menuItems.map((item) => (
-                <TableRow key={index} sx={{ '&:last-child td, &:last-child th': { border: 0 } }}>
-                  <TableCell align='right'>{item.id}</TableCell>
-                  <TableCell align="right"><Avatar src={item.images[0]}></Avatar></TableCell>
-                  <TableCell align="right">{item.name}</TableCell>
-                  <TableCell align="right">{item.price}</TableCell>
-                  <TableCell align="right">{item.status}</TableCell>
-                  <TableCell align="right">change</TableCell>
-                  <TableCell align='right'>
-                    <div className="flex justify-end gap-2">
-                      <IconButton size="small"><Edit /></IconButton>
-                      <IconButton size="small" onClick={handleDeleteFood(item.id)}><Delete /></IconButton>
-                    </div>
+              {isLoading ? (
+                <TableRow>
+                  <TableCell colSpan={7} align="center">
+                    <CircularProgress size={24} />
                   </TableCell>
                 </TableRow>
-              ))}
+              ) : menu.menuItems?.length === 0 ? (
+                <TableRow>
+                  <TableCell colSpan={7} align="center">
+                    No food items found.
+                  </TableCell>
+                </TableRow>
+              ) : (
+                menu.menuItems.map((item) => (
+                  <TableRow key={item.id} sx={{ '&:last-child td, &:last-child th': { border: 0 } }}>
+                    <TableCell align='right'>{item.id}</TableCell>
+                    <TableCell align="right">
+                      <Avatar
+                        src={item?.imagePath || ''}
+                        alt={item.name}
+                        sx={{ width: 40, height: 40, margin: 'auto' }}
+                      />
+                    </TableCell>
+                    <TableCell align="right">{item.name}</TableCell>
+                    <TableCell align="right">{item.price}</TableCell>
+                    <TableCell align="right">{item.status}</TableCell>
+                    <TableCell align="right">change</TableCell>
+                    <TableCell align='right'>
+                      <div className="flex justify-end gap-2">
+                        <IconButton size="small">
+                          <Edit />
+                        </IconButton>
+                        <IconButton size="small" onClick={() => handleDeleteFood(item.id)}>
+                          <Delete />
+                        </IconButton>
+                      </div>
+                    </TableCell>
+                  </TableRow>
+                ))
+              )}
             </TableBody>
           </Table>
         </TableContainer>
       </Box>
     </div>
-  )
+  );
 };
 
 export default MenuTable;
